* Add bucket consistency check. * Batch delete in chunk. * Add default implementation for `delete_objs`.
