+++ rust/Cargo.toml
@@ -26,7 +26,6 @@ regex = "1"
 chrono = "0"
 uuid = { version = "0.8", features = ["serde", "v4"] }
 lazy_static = "1"
-itertools = "0"
 
 # Azure
 reqwest = { version = "0", optional = true }
