+++ rust/src/storage/mod.rs
@@ -244,7 +244,7 @@ pub enum StorageError {
         /// The underlying Rusoto S3 error.
         source: rusoto_core::RusotoError<rusoto_s3::PutObjectError>,
     },
-    /// Error returned when an S3 response for a requested URI does not include body bytes.
+    /// Error representing a failure when executing an S3 DeleteObject request.
     #[cfg(any(feature = "s3", feature = "s3-rustls"))]
     #[error("Failed to delete S3 object: {source}")]
     S3Delete {
@@ -255,7 +255,7 @@ pub enum StorageError {
     /// Error representing a failure when executing an S3 DeleteObjects request.
     #[cfg(any(feature = "s3", feature = "s3-rustls"))]
     #[error("Failed to delete S3 object: {source}")]
-    S3DeleteMultiple {
+    S3BatchDelete {
         /// The underlying Rusoto S3 error.
         #[from]
         source: rusoto_core::RusotoError<rusoto_s3::DeleteObjectsError>,
@@ -439,8 +439,17 @@ pub trait StorageBackend: Send + Sync + Debug {
     /// Deletes object by `path`.
     async fn delete_obj(&self, path: &str) -> Result<(), StorageError>;
 
-    /// Deletes object by `path`.
-    async fn delete_objs(&self, paths: &[String]) -> Result<(), StorageError>;
+    /// Deletes object by `paths`.
+    async fn delete_objs(&self, paths: &[String]) -> Result<(), StorageError> {
+        for path in paths {
+            match self.delete_obj(path).await {
+                Ok(_) => continue,
+                Err(StorageError::NotFound) => continue,
+                Err(e) => return Err(e),
+            }
+        }
+        Ok(())
+    }
 }
 
 /// Dynamically construct a Storage backend trait object based on scheme for provided URI
